Extracted Fastly classes into a module
