+++ src/tests/test_buildercore_terraform.py
@@ -212,50 +212,3 @@ class TestBuildercoreTerraform(base.BaseCase):
         https://stackoverflow.com/a/16373377/91590"""
         return yaml.safe_load(terraform_template)
 
-class TestFastlyCustomVCL(base.BaseCase):
-    def test_includes_a_reference_to_itself_in_template(self):
-        snippet = terraform.FastlyCustomVCLSnippet(
-            name='do-some-magic',
-            content='...',
-            type='fetch'
-        )
-        original_main_vcl = terraform.FastlyVCL.from_string("""
-        sub vcl_fetch {
-          #FASTLY fetch
-
-          if (...) {
-            do_something_else()
-          }
-        }
-        """)
-        expected_main_vcl = terraform.FastlyVCL.from_string("""
-        sub vcl_fetch {
-          include "do-some-magic"
-
-          #FASTLY fetch
-
-          if (...) {
-            do_something_else()
-          }
-        }
-        """)
-        self.assertEqual(
-            snippet.insert_include(original_main_vcl),
-            expected_main_vcl
-        )
-
-    def test_fails_if_no_section_can_be_found(self):
-        snippet = terraform.FastlyCustomVCLSnippet(
-            name='do-some-magic',
-            content='...',
-            type='hit'
-        )
-        original_main_vcl = terraform.FastlyVCL.from_string("""
-        sub vcl_fetch {
-          ...
-        }
-        """)
-        self.assertRaises(
-            terraform.FastlyCustomVCLGenerationError,
-            lambda: snippet.insert_include(original_main_vcl),
-        )
